prevent size overflow, refs #34

+++ grails-app/controllers/docker/registry/web/RepositoryController.groovy
@@ -67,7 +67,7 @@ class RepositoryController {
     def json = restService.get("${name}/manifests/${tag}", true).json
 
     if (json.schemaVersion == 2)
-      return json.layers.collectEntries { [it.digest, it.size] }
+      return json.layers.collectEntries { [it.digest, it.size as BigInteger] }
     else {
       //fallback to manifest schema v1
       def history = json.history.v1Compatibility.collect { jsonValue ->
@@ -81,7 +81,7 @@ class RepositoryController {
       }
 
       return history.collectEntries {
-        [it.digest, it.size]
+        [it.digest, it.size as BigInteger]
       }
     }
